Add ability to expand vargs when passed as parameters
This only applies when the vargs are parameters, of course. Nothing in
Laye can return more than one value, so something like {var a, b =
vargsList} won't work as expected. In that case, a is assigned the vargs
list and b is null. a still represents a vargs list, though, so that can
be passed to a function and the effect will remain.