Update Laye
Lots of things happened here. 'each' has been changed, 'ieach' doesn't
exist anymore, just use 'each' for both. The same thing happened with
'forIEach' merging with 'forEach'. The first argument to 'forEach' is
the value, the second is the index. This way the index can be ignored if
unneeded.

Everything else is internal and doesn't affect the language directly.